Prevent kill root process

+++ src/models/processes/ProcessesManagement.java
@@ -12,6 +12,7 @@ import java.util.Vector;
 public abstract class ProcessesManagement {
   public static final int DESC = -1;
   public static final int ASC = 1;
+  public static final String ROOT_USER = "root";
   protected ArrayList<ProcessInfo> processes;
 
   public ProcessInfo find(long pid) {
@@ -70,7 +71,7 @@ public abstract class ProcessesManagement {
 
   public ProcessInfo killProcessPid(long pid) {
     ProcessInfo processInfo = find(pid);
-    if (processInfo != null) {
+    if (processInfo != null && !processInfo.getName().equals(ROOT_USER)) {
       return processInfo.kill();
     }
     return null;
